feat: login for domain module
