+++ domain/src/main/java/com/pocket/domain/dto/oauth/KakaoLoginRequest.java
@@ -1,4 +1,4 @@
-package com.pocket.domain.dto;
+package com.pocket.domain.dto.oauth;
 
 public record KakaoLoginRequest(
         String idToken
