+++ domain/src/main/java/com/pocket/domain/dto/oauth/OidcPublicKeyListResponse.java
@@ -1,4 +1,4 @@
-package com.pocket.domain.dto;
+package com.pocket.domain.dto.oauth;
 
 import java.util.List;
