+++ domain/src/main/java/com/pocket/domain/dto/user/LoginResponse.java
@@ -1,11 +1,9 @@
-package com.pocket.domain.dto;
+package com.pocket.domain.dto.user;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-public record LoginAddResponse(
-        String nickname,
-        String picture,
-        Boolean existYn,
+@Schema(description = "로그인 후 토큰 응답")
+public record LoginResponse(
         @Schema(description = "액세스 토큰", example = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ0ZXN0QGdtYWlsLmNvbSIs", defaultValue = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ0ZXN0QGdtYWlsLmNvbSIs")
         String accessToken,
         @Schema(description = "리프레시 토큰", example = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ0ZXN0QGdtYWlsLmNvbSIs", defaultValue = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ0ZXN0QGdtYWlsLmNvbSIs")
