+++ domain/src/main/java/com/pocket/domain/dto/user/UserInfoDTO.java
@@ -1,0 +1,8 @@
+package com.pocket.domain.dto.user;
+
+public record UserInfoDTO(
+        String name,
+        String email,
+        String image
+) {
+}
