+++ domain/src/main/java/com/pocket/domain/port/user/LoadUserInfoPort.java
@@ -1,0 +1,8 @@
+package com.pocket.domain.port.user;
+
+import com.pocket.domain.dto.user.UserInfoDTO;
+
+public interface LoadUserInfoPort {
+
+    UserInfoDTO loadUserInfo(String name);
+}
