+++ domain/src/main/java/com/pocket/domain/usecase/user/LoginUseCase.java
@@ -1,0 +1,8 @@
+package com.pocket.domain.usecase.user;
+
+import com.pocket.domain.dto.user.UserInfoDTO;
+
+public interface LoginUseCase {
+
+    UserInfoDTO getUserInfo(String name);
+}
